+++ tests/test_integration_crashpad.py
@@ -285,8 +285,8 @@ def wait_for_no_werfault(timeout=30.0, poll_interval=0.5):
     reason="Test covers Windows-specific crashes which can only be covered via the Crashpad WER module",
 )
 # this test currently can't run on CI because the Windows-image doesn't properly support WER, if you want to run the
-# test locally, invoke pytest with the --with_crashpad_wer option which is matched with this marker in the runtest setup
-@pytest.mark.with_crashpad_wer
+# test locally, invoke pytest with the --with_wer option which is matched with this marker in the runtest setup
+@pytest.mark.with_wer
 @pytest.mark.parametrize(
     "run_args",
     [
@@ -803,8 +803,8 @@ def test_crashpad_external_crash_reporter(cmake, httpserver, run_args):
     reason="Test covers Windows-specific crashes which can only be covered via the Crashpad WER module",
 )
 # this test currently can't run on CI because the Windows-image doesn't properly support WER, if you want to run the
-# test locally, invoke pytest with the --with_crashpad_wer option which is matched with this marker in the runtest setup
-@pytest.mark.with_crashpad_wer
+# test locally, invoke pytest with the --with_wer option which is matched with this marker in the runtest setup
+@pytest.mark.with_wer
 @pytest.mark.parametrize(
     "run_args",
     [